requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-file-share
pandas>=1.3.4
requests>=2.26.0
pathlib>=1.0.1
python-dotenv>=0.15.0
numpy>=1.20.3
pymongo>=4.0.1
dnspython>=2.1.0
python-dotenv>=0.15.0
google-api-python-client>=2.47.0
google-auth>=2.6.6
google>=3.0.0
pyodbc>=4.0.31